+++ mem_dbg/tests/test_atomics.rs
@@ -7,7 +7,7 @@ use std::sync::atomic::*;
 #[test]
 fn test_atomic_bool_in_struct() {
     #[derive(MemSize)]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Test {
         atomic: AtomicBool,
     }
@@ -23,7 +23,7 @@ fn test_atomic_bool_in_struct() {
 #[test]
 fn test_atomic_integers_in_struct() {
     #[derive(MemSize)]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Test {
         atomic_i8: AtomicI8,
         atomic_i16: AtomicI16,
@@ -63,7 +63,7 @@ fn test_atomic_integers_in_struct() {
 #[test]
 fn test_atomic_with_memdbg() {
     #[derive(MemSize, MemDbg)]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Test {
         atomic_bool: AtomicBool,
         atomic_u32: AtomicU32,