+++ mem_dbg/tests/test_mem_size.rs
@@ -17,7 +17,7 @@ use std::sync::atomic::AtomicU64;
 #[allow(dead_code)]
 #[derive(MemSize)]
 #[cfg_attr(feature = "std", derive(MemDbg))]
-#[mem_size_rec]
+#[mem_size(rec)]
 enum TestEnum {
     Unit,
     Unit2(),
@@ -45,7 +45,7 @@ fn test_vec_capacity() {
 fn test_vec_copy_or_not() {
     #[derive(Clone, MemSize)]
     #[cfg_attr(feature = "std", derive(MemDbg))]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct NewType(usize);
 
     assert_eq!(
@@ -58,7 +58,7 @@ fn test_vec_copy_or_not() {
 fn test_boxed_slice_copy_or_not() {
     #[derive(Clone, MemSize)]
     #[cfg_attr(feature = "std", derive(MemDbg))]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct NewType(usize);
 
     assert_eq!(
@@ -75,7 +75,7 @@ fn test_boxed_slice_copy_or_not() {
 fn test_slice_copy_or_not() {
     #[derive(Clone, MemSize)]
     #[cfg_attr(feature = "std", derive(MemDbg))]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct NewType(usize);
 
     assert_eq!(
@@ -94,7 +94,7 @@ fn test_slice_copy_or_not() {
 fn test_array_copy_or_not() {
     #[derive(Clone, Copy, MemSize)]
     #[cfg_attr(feature = "std", derive(MemDbg))]
-    #[mem_size_flat]
+    #[mem_size(flat)]
     struct NewType(usize);
 
     assert_eq!(
@@ -108,7 +108,7 @@ fn test_array_copy_or_not() {
 #[test]
 fn test_empty_struct() {
     #[derive(MemSize, Clone, Copy)]
-    #[mem_size_flat]
+    #[mem_size(flat)]
     struct Data {}
     let v = Data {};
     assert_eq!(v.mem_size(SizeFlags::default()), 0);
@@ -141,7 +141,7 @@ fn test_struct() {
 #[test]
 fn test_empty_tuple_struct() {
     #[derive(MemSize)]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Data();
     let v = Data();
     assert_eq!(v.mem_size(SizeFlags::default()), 0);
@@ -170,13 +170,13 @@ fn test_padding() {
     let expected = size_of::<(u8, u64)>();
     assert_eq!((0_u8, 0_u64).mem_size(SizeFlags::default()), expected);
     #[cfg_attr(feature = "derive", derive(MemSize))]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct TuplePadded((u8, u64));
     let v = TuplePadded((0, 0));
     assert_eq!(v.mem_size(SizeFlags::default()), expected);
 
     #[cfg_attr(feature = "derive", derive(MemSize))]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct StructPadded(u8, u64);
     let v = StructPadded(0, 0);
     assert_eq!(v.mem_size(SizeFlags::default()), expected);
@@ -271,7 +271,7 @@ fn test_exotic() {
     );
 
     #[cfg_attr(feature = "derive", derive(MemSize))]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     enum Data1 {
         A,
         B,
@@ -328,7 +328,7 @@ fn test_phantom() {
     struct Dummy();
     #[cfg_attr(feature = "derive", derive(MemSize))]
     #[cfg_attr(all(feature = "std", feature = "derive"), derive(MemDbg))]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Example<A>(PhantomData<A>);
 
     Example::<Dummy>(PhantomData)
@@ -368,7 +368,7 @@ fn test_array_u8() {
 #[cfg(feature = "std")]
 fn test_array_empty_struct() {
     #[derive(MemSize, MemDbg, Clone, Copy)]
-    #[mem_size_flat]
+    #[mem_size(flat)]
     struct Dummy;
     let data = [Dummy; 10];
     assert_eq!(data.mem_size(SizeFlags::default()), 0);
@@ -398,7 +398,7 @@ fn test_slice_u8() {
 #[test]
 fn test_slice_empty_struct() {
     #[derive(MemSize, MemDbg, Clone, Copy)]
-    #[mem_size_flat]
+    #[mem_size(flat)]
     struct Dummy;
     let data = [Dummy; 10].as_slice();
     assert_eq!((*data).mem_size(SizeFlags::default()), 0);