+++ mem_dbg/tests/test_mem_size_no_import.rs
@@ -3,7 +3,7 @@
 #[allow(dead_code)]
 #[cfg_attr(feature = "derive", derive(mem_dbg::MemSize))]
 #[cfg_attr(all(feature = "std", feature = "derive"), derive(mem_dbg::MemDbg))]
-#[cfg_attr(feature = "derive", mem_size_rec)]
+#[cfg_attr(feature = "derive", mem_size(rec))]
 struct MyTestStruct(i32);
 
 #[test]