+++ mem_dbg/tests/test_nonzero.rs
@@ -7,7 +7,7 @@ use std::num::*;
 #[test]
 fn test_nonzero_u8_in_struct() {
     #[derive(MemSize)]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Test {
         value: NonZeroU8,
     }
@@ -23,7 +23,7 @@ fn test_nonzero_u8_in_struct() {
 #[test]
 fn test_nonzero_integers_in_struct() {
     #[derive(MemSize)]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Test {
         nz_i8: NonZeroI8,
         nz_i16: NonZeroI16,
@@ -63,7 +63,7 @@ fn test_nonzero_integers_in_struct() {
 #[test]
 fn test_nonzero_with_memdbg() {
     #[derive(MemSize, MemDbg)]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Test {
         nz_u32: NonZeroU32,
         nz_u64: NonZeroU64,