+++ mem_dbg/tests/test_phantom.rs
@@ -7,7 +7,7 @@ use std::marker::{PhantomData, PhantomPinned};
 #[test]
 fn test_phantom_data_in_struct() {
     #[derive(MemSize)]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Test<T> {
         phantom: PhantomData<T>,
     }
@@ -23,7 +23,7 @@ fn test_phantom_data_in_struct() {
 #[test]
 fn test_phantom_pinned_in_struct() {
     #[derive(MemSize)]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Test {
         pinned: PhantomPinned,
     }
@@ -39,7 +39,7 @@ fn test_phantom_pinned_in_struct() {
 #[test]
 fn test_phantom_types_in_struct() {
     #[derive(MemSize)]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Test<T, U> {
         phantom_data: PhantomData<T>,
         phantom_pinned: PhantomPinned,
@@ -61,7 +61,7 @@ fn test_phantom_types_in_struct() {
 #[test]
 fn test_phantom_with_memdbg() {
     #[derive(MemSize, MemDbg)]
-    #[mem_size_rec]
+    #[mem_size(rec)]
     struct Test<T> {
         phantom_data: PhantomData<T>,
         phantom_pinned: PhantomPinned,